+++ src/solution.rs
@@ -1,8 +1,8 @@
 pub trait Solution {
-    fn part1(&self, _input: &str) -> Option<u64> {
+    fn part1(&self, _input: &str) -> Option<usize> {
         None
     }
-    fn part2(&self, _input: &str) -> Option<u64> {
+    fn part2(&self, _input: &str) -> Option<usize> {
         None
     }
 }
